Init theme toggler with correct position
Previously, when the page was opened, the theme toggler loaded in
the same position, no matter the theme stored in the localStorage.
This means that even if a user had the 'dark' theme stored, the
toggler would assume it was associated to the 'light' theme.
This leads to inconsistent behaviour with the toggler, since it
should ALWAYS turn the theme 'light' or 'dark' by going left or right.

Programatically clicking the toggler if the loaded theme is 'dark' fixes
the issue.