Improve color logic and code quality
* Color logic
Previously, we were directly setting the color of the elements in JS.
This, does not play well when the user changes the theme, and is
much harder to edit later on if necessary. The colors are now
defined as css variables, and the JS just changes classes according
to the action performed. The classes then set the color of the elements
based on the variables.

* Code quality
Repeated code on the runLevelX functions was organized into functions
to be re-used. This should hopefully make it much more legible.